requirements.txt
numpy
open3d
scipy
opencv-python
pandas
scikit-learn
pylsd-nova
omegaconf
pytorch_lightning
h5py
